+++ rick-and-morty-api/src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Header from "./components/Header";
 import Characters from "./components/Characters";
 import Welcome from "./components/Welcome";
