+++ rick-and-morty-api/src/reducers/index.js
@@ -1,8 +1,37 @@
-import { combineReducers } from "react-redux";
-import navReducer from "./navReducer";
-import fetchReducer from "./fetchReducer";
+import {
+  FETCH_LOADING,
+  FETCH_SUCCESS,
+  FETCH_ERROR,
+} from "../actions";
 
-export const reduccer = combineReducers({
-  navReducer,
-  fetchReducer,
-});
+const initialState = {
+  response: null,
+  error: null,
+  loading: true,
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case FETCH_LOADING:
+      return {
+        ...state,
+        loading: true,
+      };
+    case FETCH_SUCCESS:
+      return {
+        ...state,
+        loading: false,
+        response: action.payload,
+      };
+    case FETCH_ERROR:
+      return {
+        ...state,
+        loading: false,
+        error: action.payload,
+      };
+    default:
+      return state;
+  }
+};
+
+export default reducer;
